requirements.txt
catboost==1.2.2
imbalanced_learn==0.11.0
imblearn==0.0
lightgbm==4.5.0
matplotlib==3.8.2
missingno==0.5.2
numpy==2.1.0
pandas==2.2.2
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
sklearn_relief==1.0.0b2
skrebate==0.62
statsmodels==0.14.0
tabulate==0.9.0
tensorflow==2.15.0
tensorflow_macos==2.15.0
xgboost==2.1.1
